src/plugin.h: Use `int64_t' for `derive_t' and `uint64_t' for `absolute_t'.

+++ src/plugin.h
@@ -28,10 +28,10 @@
 
 #define DATA_MAX_NAME_LEN 64
 
-#define DS_TYPE_COUNTER 0
-#define DS_TYPE_GAUGE   1
+#define DS_TYPE_COUNTER  0
+#define DS_TYPE_GAUGE    1
 #define DS_TYPE_DERIVE   2
-#define DS_TYPE_ABSOLUTE   3
+#define DS_TYPE_ABSOLUTE 3
 
 #ifndef LOG_ERR
 # define LOG_ERR 3
@@ -60,15 +60,15 @@
  */
 typedef unsigned long long counter_t;
 typedef double gauge_t;
-typedef unsigned long long derive_t;
-typedef unsigned long long absolute_t;
+typedef int64_t derive_t;
+typedef uint64_t absolute_t;
 
 union value_u
 {
-	counter_t counter;
-	gauge_t   gauge;
+	counter_t  counter;
+	gauge_t    gauge;
 	derive_t   derive;
-	absolute_t   absolute;
+	absolute_t absolute;
 };
 typedef union value_u value_t;
